+++ AOC2023/Days.h
@@ -14,5 +14,5 @@ int Day3_1(vector<string>& lines);
 int Day3_2(vector<string>& lines);
 int Day4_1(vector<string>& lines);
 int Day4_2(vector<string>& lines);
-int Day5_1(vector<string>& lines);
-int Day5_2(vector<string>& lines);
+long long Day5_1(vector<string>& lines);
+long long Day5_2(vector<string>& lines);
